refactor: app flow into usecases and adapters
