+++ src/adapters/filesystem/file_copy.rs
@@ -1,0 +1,1 @@
+pub use crate::infrastructure::file_copy::copy_configured_files;
