+++ src/adapters/filesystem/mod.rs
@@ -1,0 +1,5 @@
+pub mod file_copy;
+pub mod ops;
+
+pub use file_copy::copy_configured_files;
+pub use ops::{FileSystem, RealFileSystem};
