+++ src/adapters/filesystem/ops.rs
@@ -1,0 +1,1 @@
+pub use crate::infrastructure::filesystem::{FileSystem, RealFileSystem};
