+++ src/adapters/git/worktree_lock.rs
@@ -1,0 +1,1 @@
+pub use crate::infrastructure::git::WorktreeLock;
