+++ src/adapters/mod.rs
@@ -1,0 +1,6 @@
+pub mod config;
+pub mod filesystem;
+pub mod git;
+pub mod hooks;
+pub mod shell;
+pub mod ui;
