+++ src/adapters/shell/mod.rs
@@ -1,0 +1,2 @@
+pub mod editor;
+pub mod switch_file;
