+++ src/adapters/ui/mod.rs
@@ -1,0 +1,3 @@
+pub mod dialoguer;
+
+pub use dialoguer::{DialoguerUI, MockUI, UserInterface};
