+++ src/app/mod.rs
@@ -1,0 +1,6 @@
+pub mod actions;
+pub mod menu;
+pub mod presenter;
+pub mod run;
+
+pub use run::run;
