+++ src/domain/branch.rs
@@ -1,0 +1,1 @@
+pub use crate::git_interface::{BranchInfo, TagInfo};
