+++ src/domain/mod.rs
@@ -1,0 +1,5 @@
+pub mod branch;
+pub mod paths;
+pub mod repo_context;
+pub mod validation;
+pub mod worktree;
