+++ src/domain/paths.rs
@@ -1,0 +1,3 @@
+pub use crate::usecases::create_worktree::{
+    determine_worktree_path, validate_worktree_creation, validate_worktree_location,
+};
