+++ src/domain/worktree.rs
@@ -1,0 +1,1 @@
+pub use crate::git::WorktreeInfo;
